+++ CHANGELOG.md
@@ -4,7 +4,6 @@ Pearl changelog
 
 ## [Unreleased]
 #### Added
-- Add hierarchical filters support (#126)
 
 
 #### Changed
@@ -14,6 +13,20 @@ Pearl changelog
 
 
 #### Updated
+
+
+## [0.9.2] - 2022-02-14
+#### Added
+- Add record delete api method (#103)
+
+
+## [0.9.1] - 2022-02-03
+#### Added
+- Add wait cycle for tests with index files (#144)
+
+
+#### Fixed
+- In memory index last record retrieval (#147)
 
 
 ## [0.9.0] - 2021-12-09